sửa giao diện + CRUD tuyến bay

+++ src/services/QLSanBayServices.jsx
@@ -2,7 +2,7 @@ import apiClient from "./apiClient";
 
 export const getAllSanBay = async () =>{
     try{
-        const response = await apiClient.get("/sanbay");
+        const response = await apiClient.get("/admin/dashboard/sanbay");
         return response.data;
     }
     catch (error){
@@ -11,11 +11,11 @@ export const getAllSanBay = async () =>{
     }
 }
 
-export const thongTinSanBay = async (maICAO) => { // 1. Thêm tham số maSanBay
+export const thongTinSanBay = async (maICAO) => { 
     try{
         // 2. Sử dụng maSanBay để tạo URL động
-        const response = await apiClient.get(`/sanbay/${maICAO}`); 
-        return response.data; // 3. Trả về dữ liệu từ response
+        const response = await apiClient.get(`/admin/dashboard/sanbay/${maICAO}`); 
+        return response.data; 
     } catch (error) {
         // 4. Thêm xử lý lỗi
         console.error(`Lỗi khi lấy thông tin sân bay ${maICAO}`, error);
@@ -25,7 +25,7 @@ export const thongTinSanBay = async (maICAO) => { // 1. Thêm tham số maSanBay
 
 export const addSanBay = async (sanBayData) => {
     try{
-        const response = await apiClient.post('/sanbay', sanBayData);
+        const response = await apiClient.post('/admin/dashboard/sanbay', sanBayData);
         return response.data;
     }
     catch(error){
@@ -34,9 +34,19 @@ export const addSanBay = async (sanBayData) => {
     }
 };
 
+export const updateTrangThaiSanBay = async (maSanBay, trangthai) => {
+    try {
+        const response = await apiClient.put(`/admin/dashboard/sanbay/trangthai?maSanBay=${maSanBay}&trangThai=${trangthai}`);
+        return response.data;
+    } catch (error) {
+        console.error(`Lỗi khi cập nhật trạng thái sân bay ${maSanBay}`, error);
+        throw error;
+    }
+};
+
 export const deleteSanBay = async (maSanBay) => {
     try{
-        const response = await apiClient.delete(`/sanbay/${maSanBay}`);
+        const response = await apiClient.delete(`/admin/dashboard/sanbay/${maSanBay}`);
         return response.data;
     }
     catch(error){
@@ -44,3 +54,14 @@ export const deleteSanBay = async (maSanBay) => {
         throw error;
     }
 };
+
+export const getSanBayActive = async () => {
+    try{
+        const response = await apiClient.get("/admin/dashboard/sanbay");
+        return response.data;
+    }
+    catch (error){
+        console.error("Lỗi khi lấy danh sách sân bay hoạt động", error);
+        throw error;
+    }
+}
